ci: use newer XCode (#964)
## 📜 Description

In preparation for https://github.com/actions/runner-images/issues/12195

## 💡 Motivation and Context

XCode 15.4 will be removed soon (and in fact it's already removed on
some nodes and because of this some jobs may fail). To fix this I locked
XCode version to `16.2` (I also use locally this version).

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### CI

- use `XCode 16.2` version for CI jobs;
- lock iOS version for iOS performance tests;
- update `destination` for iOS builds (to iPhone 16 Pro);
- create performance baseline for iPhone 16 Pro, iOS 18.4.

## 🤔 How Has This Been Tested?

Tested via this PR.

## 📸 Screenshots (if appropriate):

<!-- Add screenshots/video if needed -->
<!-- That would be highly appreciated if you can add how it looked
before and after your changes -->

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed