fix: 2 memory leaks (#966)
## 📜 Description

<!-- Describe your changes in detail -->

This PR fixes 3 memory leaks:
- first is connected to `ModalAttachedWatcher` keeping itself as a
listener in `EventDispatcher`, everytime the `EdgeToEdgeReactViewGroup`
sets itself to `active`. I fixed it with calling `setActive(false)` in
`onDropViewInstance` of its manager.
- second is connected to `EdgeToEdgeReactViewGroup` calling
`setupWindowDimensionsListener` and not detaching it ever. It is a
singleton listener, so I attach it once when the first
`EdgeToEdgeReactViewGroup` instance is created in `createViewInstance`
and remove it in `invalidate` of the whole managers module.
- ~~third is connected to `EdgeToEdgeReactViewGroup` also adding a
listener on `rootView` each time when calling `setupWindowInsets` and
never detaching it, so we detach it in `onDropViewInstance`.~~ <- will
be fixed later on in a separate PR

## 💡 Motivation and Context

Fix memory leaks happening e.g. on reload.
<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->

## 📢 Changelog

### Android
- Add proper detaching on listeners when views or whole modules are
destroyed

## 🤔 How Has This Been Tested?

Tested in Expensify with hitting reload.
<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, and the tests you ran
to -->
<!-- see how your change affects other areas of the code, etc. -->

## 📸 Screenshots (if appropriate):

<!-- Add screenshots/video if needed -->
<!-- That would be highly appreciated if you can add how it looked
before and after your changes -->

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed

---------

Co-authored-by: kirillzyusko <zyusko.kirik@gmail.com>