feat: keyboard state selector (#998)
## 📜 Description

Added ability to pass selector to `useKeyboardState` hook to query only
necessary updates.

## 💡 Motivation and Context

The concept of selectors should be very familiar to many RN developers,
since it's used in `redux` and `zustand` packages. I was also thinking
about optimizing the amount of re-renders for this hook, but initially
I've been thinking of using proxy (similar to what `react-hook-form` is
doing). But such approach will introduce a lot of complexity, so I
decided to keep the codebase simple and efficient, so added a concept of
selectors.

Additionally I re-worked documnetation mainly focusing on:
- adding a sample how `useKeyboardAnimation` can be used as alternative;
- making warnings/tips shorter with corresponding references;
- adding the usage for selectors concept;
- use a real-world example with `useKeyboardState` (conditional
rendering instead of changing styles).

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- allow to pass selectors to `useKeyboardState`;

### Docs

- make warning/tips shorter with corresponding links;
- added `useKeyboardState` vs `useKeyboardAnimation` code reference;
- mention how to use selectors;
- re-work example app to match real-world scenarios usage.

## 🤔 How Has This Been Tested?

Tested manually in example project.

## 📸 Screenshots (if appropriate):

|Dark theme|Light theme|
|------------|------------|

|![image](https://github.com/user-attachments/assets/11eb0c85-7d6d-43a5-9d81-0bfc072f9042)|![image](https://github.com/user-attachments/assets/693c7580-5780-4ab6-8799-43503684e075)|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed