refactor: do not rely on deep imports (#1000)
## 📜 Description

Don't rely on deep imports and provide a warning free compatibility with
RN 0.80+ (by utilizing a more official and reliable way to substitute
native module implementation).

## 💡 Motivation and Context

Starting from `react-native@0.80` deep imports were deprecated and now
if you try to use them you get a warning:

> Deep imports from the 'react-native' package are deprecated

In version `0.82` the ability to use deep imports will be completely
removed. In preparation of that RN team opened a discussion on GitHub
about those upcoming changes.

This library uses deep imports to substitute `StatusBar` implementation.
I
[opened](https://github.com/react-native-community/discussions-and-proposals/discussions/893#discussioncomment-13337514)
a thread to discuss how we can overcome it. While waiting for a proposed
solution I discovered, that we can override native modules with own
implementation - this is the case that I want to use. I want to
substitute the implementation of `StatusBar` and fallback to default
implementation if `EdgeToEdge` view is disabled.

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- remove `monkey-patch` code;

### Android

- rename `StatusBarManagerCompat` to `StatusBarManager` to match name of
original module;
- added `getConstants` method;
- `StatusBarManagerCompat` now can override existing modules;
- make `active` prop public;
- make `setActive` setter field-setter to avoid compilation conflicts;
- `StatusBarManagerCompat` now respects `active` props and fallbacks to
`StatusBarManager` implementation if `edge-to-edge` view is disabled;
- added helpers like `isEnabled`/`view` to `StatusBarManagerCompat` to
reduce code duplication.

## 🤔 How Has This Been Tested?

Tested manually on:
- Pixel 7 Pro (API 36, real device);
- Pixel 3A (API 33, emulator);

## 📸 Screenshots (if appropriate):


https://github.com/user-attachments/assets/44a7e019-446b-4c7e-903f-7d4c87882b74

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed