feat: use `keyboardAppearance` instead of global appearance (#1004)
## 📜 Description

Removed `default` as a possible value of `appearance` property.

## 💡 Motivation and Context

The thing is that for `default` a real keyboard appearance can be `dark`
or `light` depending on device preferences. In this PR I decided to
combine the logic on a native side and simplify JS logic, so for that I
removed `default` as a possible value from a native side.

The other improvement is that we can consume `appearance` property from
`useKeyboardState`. For that I had a separate listener to
`KeyboardEvents` (`keyboardWillShow` event), because we need to know the
actual field value before keyboard appears.

I removed a code that was responsible for detecting keyboard color from
`KeyboardToolbar` - now we can just use `const appearance =
useKeyboardState(state => state.appearance)` and don't go to low level
stuff with setting up listeners. Just pass selector 😎

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- change `appearance` in `useKeyboardState` hook on `willShow` event;
- use `useKeyboardState` instead of `useColorScheme`; 

### iOS

- remove `default` as possible value for `appearance`;

### Android

- remove `default` as possible value for `appearance`;

## 🤔 How Has This Been Tested?

Tested manually in example app.

## 📸 Screenshots (if appropriate):

> images taken when you set `keyboardAppearance="dark"`

|Before|After|
|-------|-----|

|![image](https://github.com/user-attachments/assets/2060afee-bc3b-4e91-a80d-239bfa13e8ec)|![image](https://github.com/user-attachments/assets/9c79313b-4de8-4ea1-89ca-f4fb1584e7ba)|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed