fix: rewrite `onTextChanged` to `onSelectionChanged` event handler in `KeyboardAwareScrollView` (#546)
## 📜 Description

Rewrite `onTextChanged` to `onSelectionChanged` event handler in
`KeyboardAwareScrollView`.

## 💡 Motivation and Context

This is the first PR that re-writes `KeyboardAwareScrollView` handlers
to `onSelectionChanged` as a main driver.

The current problem with `onTextChanged`/`onLayoutChanged` events is
that we don't know the coordinates of the caret, and in certain cases
(if view is full screen height) it may lead to unpredictable changes.

So in this PR I remove `onTextChanged` handler and replace it with
`onSelectionChanged`. I fixed e2e tests and overall have a confidence
about these changes 😊

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### E2E

- update assets;
- update selection code to make it more robust;

### JS

- use `onSelectionChanged` instead of `onTextChanged`;

## 🤔 How Has This Been Tested?

Tested in e2e tests.

## 📸 Screenshots (if appropriate):

|Before|After|
|-------|-----|

|![image](https://github.com/user-attachments/assets/d5cc152f-4e2d-42d3-a93e-a60891a2bec6)|![image](https://github.com/user-attachments/assets/4fb89683-496b-4174-8b98-4aba8a04a3bd)|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed