fix: delay delegate substitution (#1027)
## 📜 Description

Currently in my library, I have a bug that if we have text input with
autofocus, your library adds a delegate and it does it before my
library, as I do it with decorator view, which is mounted on the screen
by the next element. I think it is not so important for your library to
put its delegate( in terms of as early as possible). So I added logic
that does this on the next frame.


Advanced input mask issue:
https://github.com/IvanIhnatsiuk/react-native-advanced-input-mask/issues/128

## 💡 Motivation and Context

Allow other libraries to set their own delegates

<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### iOS

- Wrapped substituteDelegate call with DispatchQueue.main.async


## 🤔 How Has This Been Tested?

<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, and the tests you ran
to -->
<!-- see how your change affects other areas of the code, etc. -->

tested locally on iOS simulator

## 📸 Screenshots (if appropriate):

<!-- Add screenshots/video if needed -->
<!-- That would be highly appreciated if you can add how it looked
before and after your changes -->


https://github.com/user-attachments/assets/bb9e4725-d7d4-4fcb-9e28-0d94282315c0

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed