fix: conditional rendering of `OverKeyboardView` (#1040)
## 📜 Description

Fixed a problem on non-dispatching `onPress` when you render
`OverKeyboardView` conditionally.

## 💡 Motivation and Context

The fix consist of two issues

### 1️⃣ Conditional rendering

When we render element conditionally, then we can not rely on
`self.window` because it'll be `null` (view is not attached yet) and
final frame will be `null` too, so because of that touch handlers don't
work.

I fixed this problem by using dimensions from the target window (i. e.
used `topWindow` because it's always defined).

### 2️⃣ View recycling

Even though when touch problem is fixed you still can see that view only
one time 😲 The problem here is because of the fact, that view after
unmounting gets recycled, and next time, when you mount it the
condition:

```objc
if (newViewProps.visible != oldViewProps.visible) {
```

will be evaluated as `true != true` and inner statement will not be
executed. On a JS we didn't reset the `visible` prop to `false`, so
technically recycled view has the same props as just mounted one, so we
don't need to "show" it.

But technically it's incorrect, because when view was unmounted we
remove it from its window and reset inner `_visible` variable to
`false`.

After thinking a while I decided that it'll be correct to compare
`newViewProps.visible` with `_visible` to decide whether to show/hide
view or do nothing. And new approach seems to be working fine.

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/1010

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### iOS

- use `topWindow` instead of `self.window` to get correct frame;
- compare `newViewProps.visible` with `_visible` instead of
`oldViewProps.visible`;

## 🤔 How Has This Been Tested?

Tested repro code in Fabric example app.

## 📸 Screenshots (if appropriate):

|Before|After|
|-------|-----|
|<video
src="https://github.com/user-attachments/assets/59224308-e1db-4c17-bddb-e5df35e39e02">|<video
src="https://github.com/user-attachments/assets/5f67fa56-738f-438d-a43a-2a713014cc42">|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed