fix: clang format (#1060)
## 📜 Description

Fixed `clang` formatter after
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/1059
fix

## 💡 Motivation and Context

I couldn't fix the issue directly in the fork, because the fix was made
in `main` branch. Also didn't want to ask author again and again to fix
it, so decided to merge PR with failure and then apply a post-fix.

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### iOS

- run `lint-clang` to fix all the issues;

## 🤔 How Has This Been Tested?

Tested on CI.

## 📸 Screenshots (if appropriate):

<!-- Add screenshots/video if needed -->
<!-- That would be highly appreciated if you can add how it looked
before and after your changes -->

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed