fix: `keyboardLayoutGuide` availability (#1070)
## 📜 Description

Fixes compilation issues if the project supports iOS < 15.

## 💡 Motivation and Context

In #1047 I started to use `keyboardLayoutGuide`. In fact we use it only
on iOS 26, but we use a chain of transitive function invocations, so
compiler can not track it. To prevent compilation issues we need to wrap
the code that uses `keyboardLayoutGuide` in `#available` directive.

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/1069

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- wrap `keyboardLayoutGuide` usage in `#available` directive.

## 🤔 How Has This Been Tested?

<!-- Please describe in detail how you tested your changes. -->
<!-- Include details of your testing environment, and the tests you ran
to -->
<!-- see how your change affects other areas of the code, etc. -->

Tested via CI.

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed