fix: `StatusBar` race conditions (#1074)
## 📜 Description

Fixed a problem when `StatusBar` can manipulate its properties before
`EdgeToEdge` view was attached to the hierarchy.

## 💡 Motivation and Context

The original problem stems from the fact, that **sometimes** `StatusBar`
module may handle setters earlier, than `EdgeToEdge` view was attached.
As a result it adds undesired paddings and may break the keyboard
controller logic.

The first my idea was adding a condition `if (view() == null) { return
}`. However it's not very reliable, because some apps may use
`keyboard-controller` features (`OverKeyboardView`,
`KeyboardBackgroundView` etc.) without `KeyboardProvider`. In this case
the `react-native-keyboard-controller` will break all `StatusBar`
manipulations, so it's not the solution.

I started to investigate further and found out that we actually have a
race condition: view may be created but not attached to hierarchy yet.
In this case we forward call to original implementation. Here I wanted
to fix it with code that will reset all flags that were set by
StatusBar. But again the trick didn't work, because the pipeline is
next:
- view gets created/configured (I reset flags here)
- StatusBar handles its setter
- view gets attached to hierarchy.

So I had to put the code in `onAttachedToWindow` lifecycle, but I think
it would make code more complicated/harder to debug in future. So I
decided to re-work approach a little bit. Instead of using hierarchy as
source of truth I added a registry, that holds a reference. In this case
I can get more synchronous access and be sure that there is no potential
race conditions.

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/993

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### Android

- added `EdgeToEdgeViewRegistry`;
- use `EdgeToEdgeViewRegistry` for detecting `EdgeToEdge` view/its
properties.

## 🤔 How Has This Been Tested?

Tested manually on Pixel 9 PRO API 31 (emulator) with double "R" reload.
Issue was reproducible more frequently if you would try to reload during
reload. Or if you open first screen and then do a reload.

## 📸 Screenshots (if appropriate):


https://github.com/user-attachments/assets/5340c99d-2228-49b3-bf65-5d59bb462cb9

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed