fix: don't use `@RequiresApi` with old SDK versions (#1107)
## 📜 Description

Removed `@RequiresApi` annotations usage.

## 💡 Motivation and Context

When I was getting compilation issues due to usage of API which is not
available in older versions I was constantly using first option provided
by Android Studio: to add `@RequiresApi` annotation.

However it has some downsides:
- if you mark one method with this annotation - all other methods must
use the same annotation if they call this method;
- this method is kind of semantic, and if you somehow accidentally call
this method on older devices it will lead to crash.

So to make code structure a little bit cleaner I decided to remove those
annotations and switch to `if`-statements. Initially the idea was to fix
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/1104,
but turned out that [ObsoleteSdkInt] checks also `if`-statements. So for
me as a library author there is only one reliable way to fix the issue
is to suppress those warnings (if people set linter to treat warnings as
errors it will prevent the lib from being compiled)

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/1104

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### Android

- rewrite `@RequiresApi` to `if`-statements;
- suppress new `if`-statements lint warnings via `@SuppressLint`;

## 🤔 How Has This Been Tested?

Tested in example app.

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed