feat: sync layout from JS (#1213)
## 📜 Description

Added an ability to have a bi-directional communication and read/update
layout from JS on demand.

## 💡 Motivation and Context

This PR adds a command that we can dispatch from JS in order to get an
actual layout of focused input. The task sounds pretty simple, but I
anyway had some problems, and would like to describe solutions from them
here.

First of all the querying of layout is an async operation and we need to
wait for completion before running a worklet code (so that we can be
sure, that event has been actually dispatched and shared-value has
up-to-date value). But default mechanism for view commands not suppose
to return any kind of data back to the JS thread. Alternative option is
to define "module" method and return the data/completion status from
there. However in new architecture it is highly not recommended to use
such approach, because this approach relies on RN internals that can be
changed/removed from the future. So I decided to go with recommended
approach. To overcome asynchronous stuff I decided to set one-time
listener and resolve a promise when I get specific
`layoutDidSynchronize` event - this is not a perfect solution but does a
job for now and doesn't use deprecated stuff, so I'm pretty happy.

On a high level overview - we pass `update` function through the
`context`, because `context` has an access to `KeyboardControllerView`
and from there we dispatch/wait.

Also I decided to expose `assureFocusedInputVisible` method for
`KeyboardAwareScrollView`. In the past I had requirements to scroll a
little bit when validation error appears, and back to the time I just
hard-coded it like `ref.current.scrollTo(50)`, but of course this fix
didn't handle all the cases. Now it's just enough to call
`ref.current.assureFocusedInputVisible()` and enjoy pixel perfect
automatic calculations 😎

And in the end I decided to cover new functionality with e2e tests so
that we will not break them accidentally in the future 🤞

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/1208

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### Docs

- added info about new `update` method;
- added info about `assureFocusedInputVisible` method of ref;

### E2E

- cover new functionality described in
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/1208
- cover use cases when we set a focus on input that doesn't belong to
`KeyboardAwareScrollView`;

### JS

- added `synchronizeFocusedInputLayout` command for
`KeyboardControllerView`;
- added `KeyboardController::layoutDidSynchronize` event;
- expose `assureFocusedInputVisible` method for
`KeyboardAwareScrollView`;

### iOS

- implement `synchronizeFocusedInputLayout` command for paper/fabric;

### Android

- implement `synchronizeFocusedInputLayout` command for paper/fabric;

## 🤔 How Has This Been Tested?

Tested manually on iPhone 16 Pro (iOS 26.0)/Pixel 6 Pro (API 31, 35)
both paper/fabric.

## 📸 Screenshots (if appropriate):

|Android|iOS|
|-------|-----|
|<video
src="https://github.com/user-attachments/assets/42a48042-e244-4a2d-8017-606d5fabfb7e">|<video
src="https://github.com/user-attachments/assets/d72ee356-3fd6-4019-b5c6-9e09ef60608d">|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed